Don't export libusb_device
